+++ src/Typin/Typin/Directives/InteractiveDirective.cs
@@ -1,6 +1,5 @@
 ﻿namespace Typin.Directives
 {
-    using System.Diagnostics.CodeAnalysis;
     using System.Threading;
     using System.Threading.Tasks;
     using Typin.Attributes;
@@ -12,7 +11,6 @@
     /// This is useful for situations when it is necessary to execute multiple commands (since you don't have to constantly type dotnet ...).
     /// Furthermore, application context can be shared, which is useful when you have a db connection or startup takes very long.
     /// </summary>
-    [ExcludeFromCodeCoverage]
     [Directive(BuiltInDirectives.Interactive, Description = "Executs a command, then starts an interactive mode.",
                ExcludedModes = new[] { typeof(InteractiveMode) })]
     public sealed class InteractiveDirective : IPipelinedDirective