+++ src/Typin/Typin/Modes/Interactive/InteractiveModeBuilderSettings.cs
@@ -6,17 +6,17 @@
     public sealed class InteractiveModeBuilderSettings
     {
         /// <summary>
-        /// Whether to add [>], [.], and [..] directives.
+        /// Whether to add [>], [.], and [..] directives (default: true).
         /// </summary>
         public bool AddScopeDirectives { get; init; } = true;
 
         /// <summary>
-        /// Whether to add [interactive] directive.
+        /// Whether to add [interactive] directive (default: true).
         /// </summary>
         public bool AddInteractiveDirective { get; init; } = true;
 
         /// <summary>
-        /// Whether to add interactive command.
+        /// Whether to add interactive command (default: true).
         /// </summary>
         public bool AddInteractiveCommand { get; init; } = true;
     }